refactor consent request page WIP
