+++ src/features/amUI/common/RequestPageLayout/RequestPageLayout.module.css
@@ -18,6 +18,11 @@
   margin-bottom: 3px;
 }
 
+.backButton {
+  align-self: flex-start;
+  margin-bottom: var(--ds-size-2);
+}
+
 @media only screen and (max-width: 769px) {
   .requestBlock {
     padding-left: var(--ds-size-6);
