+++ src/features/amUI/consent/ConsentRequestPage/ConsentRequestPage.module.css
@@ -1,24 +1,3 @@
-.centerBlock {
-  display: block;
-  align-self: center;
-  margin-top: var(--ds-size-4);
-  max-width: 30rem;
-}
-
-.consentBlock {
-  background-color: var(--ds-color-accent-background-default);
-  padding: var(--ds-size-9) var(--ds-size-15);
-  margin-bottom: var(--ds-size-12);
-}
-.headerBlock {
-  display: flex;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: var(--ds-size-6);
-  padding-top: var(--ds-size-14);
-  margin-bottom: 3px;
-}
-
 .consentContent {
   display: flex;
   flex-direction: column;
@@ -67,17 +46,4 @@
   align-self: center;
   height: 100%;
   max-width: 27rem;
-}
-
-.backButton {
-  align-self: flex-start;
-  margin-bottom: var(--ds-size-2);
-}
-
-@media only screen and (max-width: 769px) {
-  .consentBlock {
-    padding-left: var(--ds-size-6);
-    padding-right: var(--ds-size-6);
-    margin-bottom: 6px;
-  }
 }
