+++ src/features/amUI/consent/ConsentRequestPage/ConsentRequestPage.tsx
@@ -1,30 +1,33 @@
-import React, { useMemo } from 'react';
+import React, { useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useSearchParams } from 'react-router';
-import { DsSpinner, formatDisplayName, Layout, RootProvider } from '@altinn/altinn-components';
-
+import { useNavigate, useSearchParams } from 'react-router';
+import { DsAlert, DsButton, DsHeading, DsParagraph } from '@altinn/altinn-components';
 import { useDocumentTitle } from '@/resources/hooks/useDocumentTitle';
-import { useGetConsentRequestQuery } from '@/rtk/features/consentApi';
-import { getAltinnStartPageUrl, getLogoutUrl } from '@/resources/utils/pathUtils';
-import { PartyType, useGetUserProfileQuery } from '@/rtk/features/userInfoApi';
-import { useUpdateSelectedLanguageMutation } from '@/rtk/features/settingsApi';
-
+import {
+  useApproveConsentRequestMutation,
+  useGetConsentRequestQuery,
+  useRejectConsentRequestMutation,
+} from '@/rtk/features/consentApi';
+import { PartyType } from '@/rtk/features/userInfoApi';
 import type { ConsentRequest, ProblemDetail } from '../types';
-import { getLanguage, replaceStaticMetadata } from '../utils';
+import { getLanguage, isAccepted, isExpired, isRevoked, replaceStaticMetadata } from '../utils';
 import classes from './ConsentRequestPage.module.css';
 import { ConsentRequestError } from './ConsentRequestError';
-import { ConsentRequestContent } from './ConsentRequestContent';
+import { RequestPageLayout } from '../../common/RequestPageLayout/RequestPageLayout';
+import { ConsentPath } from '@/routes/paths/consentPath';
+import { ConsentStatus } from '../components/ConsentStatus/ConsentStatus';
+import { ConsentRights } from '../components/ConsentRights/ConsentRights';
 
 export const ConsentRequestPage = () => {
   const { t, i18n } = useTranslation();
-  const [updateSelectedLanguage] = useUpdateSelectedLanguageMutation();
 
   useDocumentTitle(t('consent_request.page_title'));
   const [searchParams] = useSearchParams();
   const requestId = searchParams.get('id') ?? '';
   const language = getLanguage(i18n.language);
-
-  const { data: userData } = useGetUserProfileQuery();
+  const skipLogout = searchParams.get('skiplogout');
+  const navigate = useNavigate();
+  const [isReceiptVisible, setIsReceiptVisible] = useState<boolean>(false);
 
   const {
     data: request,
@@ -36,6 +39,63 @@ export const ConsentRequestPage = () => {
       skip: !requestId,
     },
   );
+
+  const [
+    postApproveConsent,
+    { data: approveResponse, error: approveConsentError, isLoading: isApprovingConsent },
+  ] = useApproveConsentRequestMutation();
+
+  const [
+    postRejectConsent,
+    { data: rejectResponse, error: rejectConsentError, isLoading: isRejectingConsent },
+  ] = useRejectConsentRequestMutation();
+
+  const isRequestApproved = isAccepted(request?.consentRequestEvents ?? []);
+  const isRequestRejected = isRevoked(request?.consentRequestEvents ?? []);
+  const isRequestExpired = isExpired(request?.consentRequestEvents ?? []);
+
+  const isActionButtonDisabled =
+    isApprovingConsent ||
+    isRejectingConsent ||
+    approveResponse ||
+    rejectResponse ||
+    isRequestApproved ||
+    isRequestRejected ||
+    isRequestExpired;
+
+  const approveConsent = async (): Promise<void> => {
+    if (!isActionButtonDisabled && request) {
+      try {
+        await postApproveConsent({ requestId: request.id, language }).unwrap();
+        redirectAfterAction(true);
+      } catch {
+        // Error is already tracked via approveConsentError
+      }
+    }
+  };
+
+  const rejectConsent = async (): Promise<void> => {
+    if (!isActionButtonDisabled && request) {
+      try {
+        await postRejectConsent({ requestId: request.id }).unwrap();
+        redirectAfterAction(false);
+      } catch {
+        // Error is already tracked via rejectConsentError
+      }
+    }
+  };
+
+  const redirectAfterAction = (isApproved: boolean): void => {
+    if (skipLogout && isApproved) {
+      setIsReceiptVisible(true);
+    } else if (skipLogout && !isApproved) {
+      navigate(`/${ConsentPath.Consent}/${ConsentPath.Active}`);
+    } else {
+      window.location.assign(
+        `${import.meta.env.BASE_URL}accessmanagement/api/v1/consent/request/${request?.id}/logout`,
+      );
+    }
+  };
 
   const memoizedRequest = useMemo(() => {
     return request
@@ -50,90 +110,156 @@ export const ConsentRequestPage = () => {
       : null;
   }, [request]);
 
-  const onChangeLocale = (newLocale: string) => {
-    i18n.changeLanguage(newLocale);
-    document.cookie = `selectedLanguage=${newLocale}; path=/; SameSite=Strict`;
-    updateSelectedLanguage(newLocale);
-  };
-
   const account: { name: string; type: 'person' | 'company' } = {
     name: memoizedRequest?.fromParty.name ?? '',
     type: memoizedRequest?.fromParty.type === PartyType.Person ? 'person' : 'company',
   };
 
-  return (
-    <RootProvider>
-      <Layout
-        color='neutral'
-        theme='subtle'
-        header={{
-          locale: {
-            title: t('header.locale_title'),
-            options: [
-              { label: 'Norsk (bokmål)', value: 'no_nb', checked: i18n.language === 'no_nb' },
-              { label: 'Norsk (nynorsk)', value: 'no_nn', checked: i18n.language === 'no_nn' },
-              { label: 'English', value: 'en', checked: i18n.language === 'en' },
-            ],
-            onSelect: onChangeLocale,
-          },
-          logo: {
-            href: getAltinnStartPageUrl(i18n.language),
-            title: 'Altinn',
-          },
-          currentAccount: {
-            ...account,
-            id: '',
-            icon: account,
-          },
-          globalMenu: {
-            logoutButton: {
-              label: t('header.log_out'),
-              onClick: () => {
-                const logoutUrl = getLogoutUrl();
-                window.location.assign(logoutUrl);
-              },
-            },
-            menuLabel: t('header.menu-label'),
-            backLabel: t('header.back-label'),
-            changeLabel: t('header.change-label'),
-            menu: {
-              items: [{ groupId: 'current-user', hidden: true }],
-              groups: {
-                'current-user': {
-                  title: t('header.logged_in_as_name', {
-                    name: formatDisplayName({
-                      fullName: userData?.name || '',
-                      type: 'person',
-                      reverseNameOrder: true,
-                    }),
-                  }),
-                },
-              },
-            },
-          },
-        }}
+  let heading: React.ReactNode = null;
+  let body: React.ReactNode = null;
+  const error = loadRequestError ? (
+    <ConsentRequestError
+      error={loadRequestError as { data: ProblemDetail }}
+      defaultError='consent_request.load_consent_error'
+    />
+  ) : null;
+
+  if (isReceiptVisible) {
+    heading = (
+      <DsHeading
+        level={1}
+        data-size='md'
       >
-        <div className={classes.centerBlock}>
-          {isLoadingRequest && (
-            <DsSpinner
-              aria-label={t('consent_request.loading_consent')}
-              data-size='lg'
-            />
-          )}
-          {loadRequestError && (
-            <ConsentRequestError
-              error={loadRequestError as { data: ProblemDetail }}
-              defaultError='consent_request.load_consent_error'
-            />
-          )}
-        </div>
-        {memoizedRequest && (
-          <ConsentRequestContent
-            request={memoizedRequest}
-            language={language}
+        {memoizedRequest?.isPoa
+          ? t('consent_request.receipt_header_poa')
+          : t('consent_request.receipt_header_consent')}
+      </DsHeading>
+    );
+    body = (
+      <>
+        <DsParagraph>
+          {memoizedRequest?.isPoa
+            ? t('consent_request.receipt_body_poa')
+            : t('consent_request.receipt_body_consent')}
+        </DsParagraph>
+        <DsButton
+          onClick={() => {
+            navigate(`/${ConsentPath.Consent}/${ConsentPath.Active}`, {
+              state: { createdId: requestId },
+            });
+          }}
+        >
+          {t('consent_request.receipt_back')}
+        </DsButton>
+      </>
+    );
+  } else if (memoizedRequest) {
+    heading = (
+      <>
+        <DsHeading
+          level={1}
+          data-size='md'
+        >
+          {memoizedRequest.title[language]}
+        </DsHeading>
+        {isRequestExpired && !isRequestApproved && !isRequestRejected && (
+          <ConsentStatus
+            events={memoizedRequest.consentRequestEvents}
+            isPoa={memoizedRequest.isPoa}
           />
         )}
-      </Layout>
-    </RootProvider>
+      </>
+    );
+    body = (
+      <div className={classes.consentContent}>
+        {isRequestApproved && (
+          <DsAlert data-color='warning'>
+            {memoizedRequest?.isPoa
+              ? t('consent_request.already_approved_poa')
+              : t('consent_request.already_approved')}
+          </DsAlert>
+        )}
+        {isRequestRejected && (
+          <DsAlert data-color='warning'>
+            {memoizedRequest?.isPoa
+              ? t('consent_request.already_rejected_poa')
+              : t('consent_request.already_rejected')}
+          </DsAlert>
+        )}
+        <DsParagraph className={classes.heading}>{memoizedRequest.heading[language]}</DsParagraph>
+        {memoizedRequest.consentMessage && (
+          <DsParagraph className={classes.consentMessage}>
+            {memoizedRequest.consentMessage[language]}
+          </DsParagraph>
+        )}
+        <DsParagraph className={classes.serviceIntro}>
+          {memoizedRequest.serviceIntro[language]}
+        </DsParagraph>
+        <ConsentRights
+          rights={memoizedRequest.rights}
+          language={language}
+        />
+        <DsParagraph className={classes.expiration}>
+          {memoizedRequest.expiration[language]}
+        </DsParagraph>
+        {memoizedRequest.handledBy && (
+          <DsParagraph className={classes.handledBy}>
+            {memoizedRequest.handledBy[language]}
+          </DsParagraph>
+        )}
+        {approveConsentError && (
+          <ConsentRequestError
+            error={approveConsentError as { data: ProblemDetail }}
+            defaultError={
+              memoizedRequest.isPoa
+                ? t('consent_request.approve_error_poa')
+                : t('consent_request.approve_error')
+            }
+          />
+        )}
+        {rejectConsentError && (
+          <ConsentRequestError
+            error={rejectConsentError as { data: ProblemDetail }}
+            defaultError={
+              memoizedRequest.isPoa
+                ? t('consent_request.reject_error_poa')
+                : t('consent_request.reject_error')
+            }
+          />
+        )}
+        <div className={classes.buttonRow}>
+          <DsButton
+            variant='primary'
+            aria-disabled={isActionButtonDisabled}
+            loading={isApprovingConsent}
+            onClick={approveConsent}
+          >
+            {memoizedRequest.isPoa
+              ? t('consent_request.approve_poa')
+              : t('consent_request.approve_consent')}
+          </DsButton>
+          <DsButton
+            variant='primary'
+            aria-disabled={isActionButtonDisabled}
+            loading={isRejectingConsent}
+            onClick={rejectConsent}
+          >
+            {memoizedRequest.isPoa
+              ? t('consent_request.reject_poa')
+              : t('consent_request.reject_consent')}
+          </DsButton>
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <RequestPageLayout
+      account={account}
+      isLoading={isLoadingRequest}
+      error={error}
+      heading={heading}
+      body={body}
+    />
   );
 };
